Fix reappearing items in document list
The problem was that the cell's text was not being cleared, and so, it
seems, cells would show up at the bottom of the list view when deleted.

This seems to be an artefact of how the JavaFX TreeView handles removing
cells. Cells are simply shifted around the table rather than actually
removed.

Also, if a user removes all the documents in the list, including the
last one, the document text view and word count tables will empty
themselves rather than showing the text and word counts of the last
document.